requirements.txt
# RSS Feed Parsing
feedparser==6.0.11

# HTTP Requests
requests==2.31.0

# OpenAI API
openai==1.51.2

# Date/Time handling
python-dateutil==2.8.2

# Configuration management
python-dotenv==1.0.0

# JSON handling (built-in, but listed for clarity)
# json

# Type hints support
typing-extensions==4.9.0

# Optional: For beautiful output formatting
rich==13.7.0

# Optional: For local LLM support (Ollama)
# ollama==0.1.7  # Uncomment if using local models
